test int32 type

+++ poolHashrate-web/handler/handler.go
@@ -23,7 +23,7 @@ func PoolHashrate(w http.ResponseWriter, r *http.Request) {
 	// call the backend service
 	poolHashrateClient := poolHashrate.NewPoolHashrateService("btccom.explorer.srv.poolHashrate", client.DefaultClient)
 	rsp, err := poolHashrateClient.GetPoolHashrate(context.TODO(), &poolHashrate.Request{
-		PoolID:   request["poolId"].(int32),
+		PoolID:   int32(request["poolId"].(float64)),
 	})
 	if err != nil {
 		http.Error(w, err.Error(), 500)
